requirements.txt
functions-framework==3.*
pandas>=1.2.2
google-cloud-storage>=1.36.0